update readme
change r-tree to grid. I did not find a straight and easy way to slightly change b-tree to facilitate its use also for r-tree. The split and merge are different, i.e. merge requires re-insertion of those items that belong to removed node, while split requires to re-organize items between former and new node in a way that minimizes total area (there are couple of heuristics to do that, to name one is quadratic algorithm). So it will require significantly alter two major pieces of b-tree. To make it worse, key of b-tree is designed to be singleton and comparable object, but r-tree has bounding box rectangle (for 2D). And, search is drastically different, since it traverses more than one branch. To conclude, it is not a trivial to convert b-tree to r-tree, I would say the reverse would be more feasible task (i.e. if I would already have r-tree and needed to convert it into b-tree,, since r-tree seems more generic algorithm).